refactor: (#22) 리뷰 반영
