+++ dms-application/src/main/kotlin/team/aliens/dms/domain/student/usecase/CheckEmailDuplicateUseCase.kt
@@ -1,6 +1,6 @@
 package team.aliens.dms.domain.student.usecase
 
-import team.aliens.dms.domain.student.exception.StudentEmailExistsException
+import team.aliens.dms.domain.user.exception.StudentEmailExistsException
 import team.aliens.dms.domain.student.spi.StudentQueryUserPort
 import team.aliens.dms.global.annotation.ReadOnlyUseCase
