+++ dms-application/src/test/kotlin/team/aliens/dms/student/usecase/CheckEmailDuplicateUseCaseTest.kt
@@ -8,7 +8,7 @@ import org.junit.jupiter.api.extension.ExtendWith
 import org.mockito.BDDMockito.given
 import org.springframework.boot.test.mock.mockito.MockBean
 import org.springframework.test.context.junit.jupiter.SpringExtension
-import team.aliens.dms.domain.student.exception.StudentEmailExistsException
+import team.aliens.dms.domain.user.exception.StudentEmailExistsException
 import team.aliens.dms.domain.student.spi.StudentQueryUserPort
 import team.aliens.dms.domain.student.usecase.CheckEmailDuplicateUseCase
 
@@ -28,21 +28,21 @@ class CheckEmailDuplicateUseCaseTest {
     }
 
     @Test
-    fun `이메일 중복`() {
-        given(studentQueryUserPort.existsByEmail(email))
-            .willReturn(true)
-
-        assertThrows<StudentEmailExistsException> {
-            checkEmailDuplicateUseCase.execute(email)
-        }
-    }
-
-    @Test
     fun `이메일 중복 없음`() {
         given(studentQueryUserPort.existsByEmail(email))
             .willReturn(false)
 
         assertDoesNotThrow {
+            checkEmailDuplicateUseCase.execute(email)
+        }
+    }
+
+    @Test
+    fun `이메일 중복`() {
+        given(studentQueryUserPort.existsByEmail(email))
+            .willReturn(true)
+
+        assertThrows<StudentEmailExistsException> {
             checkEmailDuplicateUseCase.execute(email)
         }
     }
