+++ dms-domain/src/main/kotlin/team/aliens/dms/domain/user/error/UserErrorCode.kt
@@ -7,7 +7,9 @@ enum class UserErrorCode(
     private val message: String
 ) : ErrorProperty {
 
-    USER_NOT_FOUND(404, "User Not Found");
+    USER_NOT_FOUND(404, "User Not Found"),
+    STUDENT_EMAIL_EXISTS(409, "Student Email Exists")
+    ;
 
     override fun status(): Int = status
     override fun message(): String = message
