+++ dms-infrastructure/src/main/kotlin/team/aliens/dms/persistence/user/repository/UserJpaRepository.kt
@@ -7,5 +7,5 @@ import java.util.UUID
 
 @Repository
 interface UserJpaRepository : CrudRepository<UserJpaEntity, UUID> {
-    fun findByEmail(email: String): UserJpaEntity?
+    fun existsByEmail(email: String): UserJpaEntity?
 }
