+++ dms-presentation/src/main/kotlin/team/aliens/dms/student/StudentWebAdapter.kt
@@ -1,11 +1,11 @@
 package team.aliens.dms.student
 
+import org.springframework.validation.annotation.Validated
 import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.RequestMapping
 import org.springframework.web.bind.annotation.RequestParam
 import org.springframework.web.bind.annotation.RestController
 import team.aliens.dms.domain.student.usecase.CheckEmailDuplicateUseCase
-import javax.validation.Valid
 import javax.validation.constraints.NotBlank
 
 @RequestMapping("/students")
@@ -15,7 +15,7 @@ class StudentWebAdapter(
 ) {
 
     @GetMapping("/email/duplication")
-    fun checkDuplicateEmail(@RequestParam @Valid @NotBlank email: String) {
+    fun checkDuplicateEmail(@RequestParam @Validated @NotBlank email: String) {
         checkEmailDuplicateUseCase.execute(email)
     }
 }
